# openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: FusionBikes — API móvil v1
  version: 1.0.0
  description: |
    Contrato congelado para la app móvil (React Native). Fuente de verdad para ambos lados
    (app y backend) — ver docs/superpowers/plans/plan-maestro-v2.md para el plan vigente.
    La autenticación, dispositivos, notificaciones e inbox operativo de Claims están
    implementados bajo /api/v1. Los demás paths se incorporan por entregas verticales.

    Convenciones:
    - Todas las fechas en ISO 8601 UTC.
    - Todos los IDs de recurso son string (aunque el backend interno use INTEGER autoincrement).
    - El shape de error es uniforme en todos los endpoints (ver `Error` en components/schemas).
    - Paginación por cursor en listados (`cursor`/`next_cursor`), no por offset.
servers:
  - url: https://herramientas.fusionbikes.com.ar/api
    description: Producción (cuando exista)
  - url: http://localhost:3001/api
    description: Local (DISABLE_CRONS=true, nunca contra data/fusion.sqlite real)
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login con usuario/contraseña existentes (reusa la tabla `users`)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              description: Se debe enviar device_id o, en el primer registro, platform y push_token. device_id no puede combinarse con platform, push_token ni device_name.
              oneOf:
                - required: [device_id]
                  not:
                    anyOf:
                      - required: [platform]
                      - required: [push_token]
                      - required: [device_name]
                - required: [platform, push_token]
                  not:
                    required: [device_id]
              properties:
                username: { type: string, example: "juan" }
                password: { type: string, format: password, example: "••••••••" }
                device_id: { type: string, description: "ID de dispositivo ya registrado" }
                platform: { type: string, enum: [ios, android, web], description: "Usar junto con push_token en el primer login" }
                push_token: { type: string, description: "Token APNs/FCM para registrar el dispositivo en el primer login" }
                device_name: { type: string }
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
              example:
                access_token: "eyJhbGciOi..."
                refresh_token: "8f14e45f-ceea-467e-..."
                expires_in: 900
                device_id: "12"
                user: { id: "12", username: "juan", roles: ["deposito"] }
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: "credenciales_invalidas", message: "Usuario o contraseña incorrectos" } }
        '422':
          description: Body inválido o dispositivo ausente/no perteneciente al usuario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Demasiados intentos fallidos desde este usuario/IP
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Renueva el access_token usando un refresh_token vigente
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401':
          description: refresh_token inválido, expirado o revocado (ej. dispositivo eliminado)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: "refresh_revocado", message: "La sesión de este dispositivo fue revocada" } }
        '422':
          description: Falta refresh_token en el body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoca el refresh_token actual (logout de este dispositivo, no de todos)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Logout exitoso
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean, example: true } } }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /inbox:
    get:
      tags: [inbox]
      summary: Lista trabajo operativo visible para el usuario
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [unread, read, resolved, archived] }
        - name: channel
          in: query
          schema: { type: string, example: ml }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Ítems de inbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/InboxItem' } }
                  next_cursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /inbox/{id}:
    get:
      tags: [inbox]
      summary: Obtiene un ítem de inbox
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200':
          description: Ítem
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /inbox/{id}/read:
    post:
      tags: [inbox]
      summary: Marca un ítem como leído
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200': { description: Ítem actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/InboxItem' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /inbox/{id}/resolve:
    post:
      tags: [inbox]
      summary: Resuelve un ítem con control optimista de versión
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties: { version: { type: integer, minimum: 1 } }
      responses:
        '200': { description: Ítem resuelto, content: { application/json: { schema: { $ref: '#/components/schemas/InboxItem' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Conflicto de versión }
  /inbox/{id}/claim:
    post:
      tags: [inbox]
      summary: Toma el ítem y crea una notificación lógica para el usuario
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses:
        '200': { description: Ítem asignado, content: { application/json: { schema: { $ref: '#/components/schemas/InboxItem' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: El ítem ya está asignado a otro usuario }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Sin sesión válida o sesión expirada
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { ok: false, error: "No autenticado" }
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: "no_encontrado", message: "El recurso solicitado no existe" } }
    UnprocessableEntity:
      description: Request inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: Error interno no esperado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: "error_interno", message: "Ocurrió un error inesperado, reintentá" } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "Identificador estable, no cambia entre versiones" }
            message: { type: string, description: "Texto legible, en español, para mostrar o loguear" }
    AuthTokens:
      type: object
      required: [access_token, refresh_token, expires_in, device_id, user]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, description: "Segundos de vida del access_token" }
        device_id: { type: string, description: "Dispositivo al que queda ligada la sesión" }
        user: { $ref: '#/components/schemas/Me' }
    Me:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        roles: { type: array, items: { type: string } }
        permisos: { type: array, items: { type: string } }
...
    InboxItem:
      type: object
      required: [id, event_id, channel, status, version, created_at, updated_at]
      properties:
        id: { type: string }
        event_id: { type: string }
        channel: { type: string }
        resource_id: { type: string, nullable: true }
        title: { type: string }
        preview: { type: string, nullable: true }
        status: { type: string, enum: [unread, read, resolved, archived] }
        version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...